+++ lib_vos/vos_model/flow_align/src/flow_align_cuda_kernel.cu
@@ -12,7 +12,7 @@ extern "C" {
     i += blockDim.x * gridDim.x)
 
 
-__global__ void FlowAlignForward_kernel(const int nthreads,const  int height, const int width, const int channels, const double* bottom, const double* flow,double* top)
+__global__ void FlowAlignForward_kernel(const int nthreads,const  int height, const int width, const int channels, const float* bottom, const float* flow,float* top)
 {
 
   CUDA_1D_KERNEL_LOOP(index, nthreads)
@@ -25,11 +25,11 @@ __global__ void FlowAlignForward_kernel(const int nthreads,const  int height, co
       // get flow
       int ind_flow_x = w+h*width+n*height*width*2;
       int ind_flow_y = w+h*width+width*height+n*height*width*2;
-      double flo_x = flow[ind_flow_x];
-      double flo_y = flow[ind_flow_y];
+      float flo_x = flow[ind_flow_x];
+      float flo_y = flow[ind_flow_y];
       // get bilinear positions
-      double w_flo = w+flo_x;
-      double h_flo = h+flo_y;
+      float w_flo = w+flo_x;
+      float h_flo = h+flo_y;
       if ( h_flo<0 || h_flo>height-1 || w_flo<0 || w_flo>width-1 )
       {
         //outof image, we set it to 0.
@@ -39,8 +39,8 @@ __global__ void FlowAlignForward_kernel(const int nthreads,const  int height, co
         int h_start = floor(h_flo);
         int w_start = floor(w_flo);
         int nc_start = n*height*width*channels+c*height*width;
-        double h_ratio = h_flo - (double)h_start;
-        double w_ratio = w_flo - (double)w_start;
+        float h_ratio = h_flo - (float)h_start;
+        float w_ratio = w_flo - (float)w_start;
         int upleft = nc_start+w_start+width*h_start;
         int upright = upleft+1;
         int downleft = upleft+width;
@@ -54,7 +54,7 @@ __global__ void FlowAlignForward_kernel(const int nthreads,const  int height, co
   }
 }
 
-__global__ void FlowAlignBackward_kernel(const int nthreads, const int height, const int width, const int channels, const double* topdiff,const double* bottom, const double*flow, double* bottomdiff, double* flowdiff)
+__global__ void FlowAlignBackward_kernel(const int nthreads, const int height, const int width, const int channels, const float* topdiff,const float* bottom, const float*flow, float* bottomdiff, float* flowdiff)
 {
   CUDA_1D_KERNEL_LOOP(index, nthreads)
   {
@@ -66,11 +66,11 @@ __global__ void FlowAlignBackward_kernel(const int nthreads, const int height, c
       // get flow
       int ind_flow_x = w+h*width+n*height*width*2;
       int ind_flow_y = w+h*width+width*height+n*height*width*2;
-      double flo_x = flow[ind_flow_x];
-      double flo_y = flow[ind_flow_y];
+      float flo_x = flow[ind_flow_x];
+      float flo_y = flow[ind_flow_y];
       // get bilinear positions
-      double w_flo = w+flo_x;
-      double h_flo = h+flo_y;
+      float w_flo = w+flo_x;
+      float h_flo = h+flo_y;
       if ( h_flo<0 || h_flo>height-1 || w_flo<0 || w_flo>width-1 )
       {
         //outof image, no grad propagated.
@@ -81,8 +81,8 @@ __global__ void FlowAlignBackward_kernel(const int nthreads, const int height, c
         int h_start = floor(h_flo);
         int w_start = floor(w_flo);
         int nc_start = n*height*width*channels+c*height*width;
-        double h_ratio = h_flo - (double)h_start;
-        double w_ratio = w_flo - (double)w_start;
+        float h_ratio = h_flo - (float)h_start;
+        float w_ratio = w_flo - (float)w_start;
         int upleft = nc_start+w_start+width*h_start;
         int upright = upleft+1;
         int downleft = upleft+width;
@@ -95,10 +95,10 @@ __global__ void FlowAlignBackward_kernel(const int nthreads, const int height, c
         atomicAdd(bottomdiff+downright, topdiff[index] * (h_ratio) * (w_ratio));
 
         //approximate flow diff
-        double f1 = bottom[upleft];
-        double f2 = bottom[upright];
-        double f3 = bottom[downleft];
-        double f4 = bottom[downright];
+        float f1 = bottom[upleft];
+        float f2 = bottom[upright];
+        float f3 = bottom[downleft];
+        float f4 = bottom[downright];
         //recalculate top rather than saving it temporarily in the layer.
         /*
         top[index] = bottom[upleft] * (1.-h_ratio) * (1.-w_ratio)
@@ -107,16 +107,16 @@ __global__ void FlowAlignBackward_kernel(const int nthreads, const int height, c
                          + bottom[downright] * (h_ratio) * (w_ratio);
                          */        
         //calculate dx
-        double dx = -f1*(1.-h_ratio) + f2*(1.-h_ratio) - f3*(h_ratio) + f4*(h_ratio);
+        float dx = -f1*(1.-h_ratio) + f2*(1.-h_ratio) - f3*(h_ratio) + f4*(h_ratio);
         //calculate dy
-        double dy = -f1*(1.-w_ratio) - f2*(w_ratio) + f3*(1.-w_ratio) + f4*(w_ratio);
+        float dy = -f1*(1.-w_ratio) - f2*(w_ratio) + f3*(1.-w_ratio) + f4*(w_ratio);
         atomicAdd(flowdiff+ind_flow_x, topdiff[index]*dx);
         atomicAdd(flowdiff+ind_flow_y, topdiff[index]*dy);
       }
   }
 }
 
-int FlowAlignForward(const int batches, const int height, const int width, const int channels, const double* bottom, const double* flow,double* top, cudaStream_t stream)
+int FlowAlignForward(const int batches, const int height, const int width, const int channels, const float* bottom, const float* flow,float* top, cudaStream_t stream)
 {
   const int kThreadPerBlock = 512;
   const int nthreads = batches*height*width*channels;
@@ -135,7 +135,7 @@ int FlowAlignForward(const int batches, const int height, const int width, const
 }
 
 
-int FlowAlignBackward(const int batches, const int height, const int width, const int channels, const double* topdiff, const double*bottom, const double*flow, double* bottomdiff, double* flowdiff, cudaStream_t stream)
+int FlowAlignBackward(const int batches, const int height, const int width, const int channels, const float* topdiff, const float*bottom, const float*flow, float* bottomdiff, float* flowdiff, cudaStream_t stream)
 {
   const int kThreadPerBlock = 512;
   const int nthreads = batches*height*width*channels;
