+++ lib_vos/vos_model/flow_align/src/flow_align_cuda_kernel.h
@@ -5,13 +5,13 @@
 extern "C"{
 #endif
 
-__global__ void FlowAlignForward_kernel(const int nthreads,const  int height, const int width, const int channels, const double* bottom, const double* flow,double* top);
+__global__ void FlowAlignForward_kernel(const int nthreads,const  int height, const int width, const int channels, const float* bottom, const float* flow,float* top);
 
-int FlowAlignForward(const int batches,const  int height,const  int width,const  int channels, const double* bottom, const double* flow,double* top, cudaStream_t stream);
+int FlowAlignForward(const int batches,const  int height,const  int width,const  int channels, const float* bottom, const float* flow,float* top, cudaStream_t stream);
 
-__global__ void FlowAlignBackward_kernel(const int nthreads, const int height, const int width, const int channels, const double* topdiff, const double*bottom,const double*flow, double* bottomdiff, double* flowdiff);
+__global__ void FlowAlignBackward_kernel(const int nthreads, const int height, const int width, const int channels, const float* topdiff, const float*bottom,const float*flow, float* bottomdiff, float* flowdiff);
 
-int FlowAlignBackward(const int batches, const int height, const int width, const int channels, const double* topdiff, const double*bottom,const double*flow, double* bottomdiff, double* flowdiff, cudaStream_t stream);
+int FlowAlignBackward(const int batches, const int height, const int width, const int channels, const float* topdiff, const float*bottom,const float*flow, float* bottomdiff, float* flowdiff, cudaStream_t stream);
 
 
 #ifdef __cplusplus
